Hitbox work
I realized my initialization for the vector was wrong, suprising that everything else was working as intended...

I completed the distance to line function, so now we can work on circle vs rectangle hitboxes. I also realized that I am not plugging in the scalar from pixels to meters for the vector class, so gonna fix next.