Construction
I have finshed the method dist line segment, so now I have to put it together to determine if a hitbox is touching another hitbox. The hiccup is that I sense a lot of extra processing power that comes with testing each segment against another, which I currently do not know how to solve. I added a few methods to the Vector class, and I will most likely need a math library soon of my own, although I have no idea how to achieve that.